+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,25 +1,16 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
 import com.techcourse.config.DataSourceConfig;
 import com.techcourse.domain.User;
-import java.sql.Connection;
-import java.sql.SQLException;
-import javax.sql.DataSource;
 
 public class TxUserService implements UserService {
 
     private final UserService userService;
-    private final DataSource dataSource;
+    private final TransactionManager transactionManager;
 
-    public TxUserService(UserService userService, DataSource dataSource) {
+    public TxUserService(UserService userService, TransactionManager transactionManager) {
         this.userService = userService;
-        this.dataSource = dataSource;
-    }
-
-    public TxUserService(UserService userService) {
-        this(userService, DataSourceConfig.getInstance());
+        this.transactionManager = transactionManager;
     }
 
     @Override
@@ -29,32 +20,12 @@ public class TxUserService implements UserService {
 
     @Override
     public void insert(User user) {
-        executeWithAutoCommitFalse(() -> userService.insert(user));
+        transactionManager.execute(()->userService.insert(user));
     }
 
     @Override
     public void changePassword(long id, String newPassword, String createBy) {
-        executeWithAutoCommitFalse(() -> userService.changePassword(id, newPassword, createBy));
+        transactionManager.execute(() -> userService.changePassword(id, newPassword, createBy));
     }
 
-    private void executeWithAutoCommitFalse(Runnable runnable) {
-        Connection connection = DataSourceUtils.getConnection(dataSource);
-        try {
-            connection.setAutoCommit(false);
-            runnable.run();
-            connection.commit();
-        } catch (Exception e) {
-            rollback(connection);
-        } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
-    }
-
-    private void rollback(Connection connection) {
-        try {
-            connection.rollback();
-        } catch (SQLException e) {
-            throw new DataAccessException("datasource접근 과정에서 예외가 발생했습니다.", e);
-        }
-    }
 }
